requirements.txt
Flask
gunicorn
json5
google-cloud-aiplatform
google-api-python-client
requests
PyPDF2
pdfminer
pdfminer.high_level
pdfplumber